Address guided review cache follow-ups
