+++ brokk-shared/src/main/java/ai/brokk/analyzer/cache/JsTsAnalyzerCache.java
@@ -1,20 +1,18 @@
 package ai.brokk.analyzer.cache;
 
+import ai.brokk.analyzer.JsTsAnalyzer;
 import ai.brokk.analyzer.ProjectFile;
+import ai.brokk.analyzer.javascript.JsTsModuleResolution;
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
 /** JS/TS-specific analyzer caches for module and path resolution. */
 public final class JsTsAnalyzerCache extends AnalyzerCache {
-    private static final List<String> KNOWN_EXTENSIONS = List.of(".js", ".jsx", ".ts", ".tsx");
-
-    private final Cache<ModulePathKey, ResolutionOutcome> moduleResolutionCache;
+    private final Cache<ModulePathKey, JsTsAnalyzer.ResolutionOutcome> moduleResolutionCache;
     private final Cache<ModulePathFromBaseKey, Optional<ProjectFile>> moduleResolutionFromBaseCache;
 
     public JsTsAnalyzerCache() {
@@ -63,7 +61,7 @@ public final class JsTsAnalyzerCache extends AnalyzerCache {
         });
     }
 
-    public Cache<ModulePathKey, ResolutionOutcome> moduleResolutionCache() {
+    public Cache<ModulePathKey, JsTsAnalyzer.ResolutionOutcome> moduleResolutionCache() {
         return moduleResolutionCache;
     }
 
@@ -75,47 +73,8 @@ public final class JsTsAnalyzerCache extends AnalyzerCache {
 
     public record ModulePathFromBaseKey(Path baseDir, String modulePath) {}
 
-    public record ResolutionOutcome(Optional<ProjectFile> resolved, Optional<String> externalFrontier) {
-        public static ResolutionOutcome resolved(ProjectFile file) {
-            return new ResolutionOutcome(Optional.of(file), Optional.empty());
-        }
-
-        public static ResolutionOutcome external(String specifier) {
-            return new ResolutionOutcome(Optional.empty(), Optional.of(specifier));
-        }
-
-        public static ResolutionOutcome empty() {
-            return new ResolutionOutcome(Optional.empty(), Optional.empty());
-        }
-    }
-
     private static boolean changedFilesCouldAffect(ModulePathFromBaseKey key, Set<Path> changedAbsPaths) {
-        return candidatePaths(key.baseDir(), key.modulePath()).stream().anyMatch(changedAbsPaths::contains);
-    }
-
-    public static List<Path> candidatePaths(Path baseDir, String modulePath) {
-        Path resolvedPath = baseDir.resolve(modulePath).normalize();
-        String fileName = resolvedPath.getFileName().toString();
-        var candidates = new ArrayList<Path>();
-
-        if (KNOWN_EXTENSIONS.stream().anyMatch(fileName::endsWith)) {
-            candidates.add(resolvedPath);
-        }
-        String baseName = fileName;
-        for (String ext : KNOWN_EXTENSIONS) {
-            if (baseName.endsWith(ext)) {
-                baseName = baseName.substring(0, baseName.length() - ext.length());
-                break;
-            }
-        }
-        Path basePath = resolvedPath.resolveSibling(baseName);
-        candidates.add(basePath);
-        for (String ext : KNOWN_EXTENSIONS) {
-            candidates.add(basePath.resolveSibling(baseName + ext));
-        }
-        KNOWN_EXTENSIONS.stream()
-                .map(ext -> resolvedPath.resolve("index" + ext))
-                .forEach(candidates::add);
-        return List.copyOf(candidates);
+        return JsTsModuleResolution.candidatePaths(key.baseDir(), key.modulePath()).stream()
+                .anyMatch(changedAbsPaths::contains);
     }
 }
